pyproject.toml
[tool.poetry]
name = "blast-beat-pic"
version = "0.1.0"
description = ""
authors = ["Sujit Kamthe <sujitk@sahaj.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^11.0.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
